+++ test/test_helper.exs
@@ -1,5 +1,5 @@
 Code.require_file("helpers/connection.exs", __DIR__)
-Code.require_file("helpers/line_connection.exs", __DIR__)
+Code.require_file("helpers/json_connection.exs", __DIR__)
 
 Code.require_file("helpers/anon_connection.exs", __DIR__)
 Code.require_file("helpers/guest_connection.exs", __DIR__)
@@ -14,7 +14,7 @@ alias Instream.TestHelpers
 
 [
   TestHelpers.Connection.child_spec,
-  TestHelpers.LineConnection.child_spec,
+  TestHelpers.JSONConnection.child_spec,
 
   TestHelpers.AnonConnection.child_spec,
   TestHelpers.GuestConnection.child_spec,